--- backend/src/utils/getenv.ts
@@ -1,11 +1,0 @@
-import 'dotenv/config'
-
-export const getEnv = (varName: string): string => {
-  const value = process.env[varName]
-  if (!value) {
-    throw new Error(`Variable ${varName} undefined`)
-  }
-  return value
-}
-
-
